Modify the calling order: Dataset -> Agent -> Model
